Add basic layout and styles.
