+++ src/components/Layout/Layout.module.css
@@ -1,0 +1,10 @@
+.wrapper {
+  flex: 1 0 auto;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: var(--max-width);
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
