+++ src/components/content-mapper.tsx
