+++ src/components/node-article.tsx
